pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexusai"
version = "0.2.0"
description = "OpenClaw skill injection and RL training — one-click deployment"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "tiktoken",
]

[project.optional-dependencies]
# Full RL training (requires Tinker cloud access)
rl = [
    "torch",
    "transformers>=4.51.1",
    "tinker",
]
# Embedding-based skill retrieval
embedding = [
    "numpy",
    "sentence-transformers",
]
# Skill evolution via LLM
evolve = [
    "openai",
]
# W&B training metrics
wandb = [
    "wandb",
]
# Everything
all = [
    "nexusai[rl,embedding,evolve,wandb]",
]

[project.scripts]
nexusai = "nexusai.cli:nexusai"

[tool.setuptools.packages.find]
where = ["."]
include = ["nexusai*"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E731", "F403", "F405"]
